+++ src/redux/api/baseApi.ts
@@ -1,9 +1,11 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 
+const BASE_URL = 'https://ya-praktikum.tech/api/v2';
+
 export const baseApi = createApi({
   reducerPath: 'baseApi',
   baseQuery: fetchBaseQuery({
-    baseUrl: 'https://ya-praktikum.tech/api/v2',
+    baseUrl: BASE_URL,
     credentials: 'include',
   }),
   endpoints: () => ({}),
